% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDST.r
\name{runDST}
\alias{runDST}
\title{Web-interface version of the DST}
\usage{
runDST()
}
\description{
Runs the DST in a web browser using R/Shiny technology
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
